split generated op.rs into op.rs and op_code.rs
