+++ crates/ir2/src/op.rs
@@ -23,24 +23,8 @@ impl Clone for Op {
     }
 }
 
-impl Copy for OpCode {}
-impl Clone for OpCode {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-impl From<OpCode> for u16 {
-    fn from(code: OpCode) -> Self {
-        code as u16
-    }
-}
-
 #[test]
 fn op_size_of_and_alignment() {
     assert_eq!(core::mem::size_of::<Op>(), 24);
     assert_eq!(core::mem::align_of::<Op>(), 8);
 }
-
-/// Indicated an invalid `u16` value for an [`OpCode`].
-#[derive(Debug, Copy, Clone)]
-pub struct InvalidOpCode;
